rename service files
